Web App that display list of products
C# Core Web Application Activity 3a Database app with Mockaroo Data

We build a C# Web application in Visual Studio that performs database
operations Create Read Update and Delete.
In this example we build a databse and fill it with fake data form Mockaroo.
We used Faker which is a fake data generator for .NET
Created a static list of 104 items to be displayed in ./product